tui: move the Mirrors screen into tui/mirror.py
Forty-six definitions that nothing outside them used, so `settings.py` is the
only caller and `screens.py` no longer imports any of it.
