+++ tests/unit/test_compat.py
@@ -763,7 +763,7 @@ def test_a_mirror_with_no_ipv6_is_refused_on_an_ipv6_only_machine() -> None:
     from gentoo_install.i18n import Catalog
     from gentoo_install.model import mirrors
     from gentoo_install.tui.context import Context
-    from gentoo_install.tui.screens import _unreachable_here
+    from gentoo_install.tui.mirror import _unreachable_here
 
     def machine(ipv4: bool) -> Context:
         return Context(
@@ -1021,7 +1021,7 @@ def test_the_subarch_table_holds_what_the_menu_offers() -> None:
     validates it read the same set, or the menu offers what validation
     refuses."""
     from gentoo_install.model.compat import BINHOST_SUBARCHS
-    from gentoo_install.tui.screens import BINHOSTS
+    from gentoo_install.tui.mirror import BINHOSTS
 
     offered = {subarch for (_, subarch), _ in BINHOSTS}
     assert offered <= BINHOST_SUBARCHS, offered - BINHOST_SUBARCHS
